Set an id on index definitions for use later in the process
3 tests are failing out of my current selections, map reduces are being marked as non-stale when they have not yet finished indexing gosh darnit

Wholly broken code, but you can see where I'm going with this

Compiles apart from the studio, really need to work out where to use ints and where to use strings having now mullered this

Now compiles, and have renamed Name -> Id to make it obvious, nearly to the point of running tests again

Remove compiled generators ala isue #1269

Tests all failing because storage isn't using ids yet, so time to force that

Switched core storage to using ints, now need to spread that among the other core storage classes (mapreduce) and get the tests building again - they use the names to talk to storage and of course, names aren't used in storage any more

Mapped storage converted across to id usage

Converted staleness across to use index ids

Everything apart from studio builds, and most of the storage tests pass, just not the indexing tests

Managed storage doesn't throw a wobbly on integers, but it is expecting strings on read calls

Fixed a couple of snafus, now pondering on why indexes keep coming back as stale - is indexing even taking place I JUST DON'T KNOW

Hacky fix that might not work, silverlight client downloads definitions AND stats and combines the two lists

reduce breaking change by renaming PublicName back to Name now we are using Ids everywhere

Lookup tables for name -> id

Switched to using integer based ids in actual storage

Check for index not existing before deleting

What we need is to rename the column and copy data around, stashing this so I can create a database in the old version to test with

Start by creating the new one then I guess, then we'll rename outside the Tx

Made the bold move of giving the configuration to the updaters because then they can do more things

Some of the Esent stuff written, and the rest of the path is pretty layed out

My cunning plan to delete the old column is for nought, as renaming apparently also updates indexes

Gonna need to do it on a table by table basis, bugger

We need two different strategies - one for where we can do an in-place copy and deal with an auto-increment primary key, and one where we have to copy and re-create the table because we have a natural key

Migration works

+++ Raven.Database/DocumentDatabase.cs
@@ -411,10 +411,10 @@ namespace Raven.Database
                     {
 	                    try
 	                    {
-		                    index.LastQueryTimestamp = IndexStorage.GetLastQueryTime(index.Name);
-		                    index.Performance = IndexStorage.GetIndexingPerformance(index.Name);
-		                    index.IsOnRam = IndexStorage.IndexOnRam(index.Name);
-		                    var indexDefinition = IndexDefinitionStorage.GetIndexDefinition(index.Name);
+		                    var indexDefinition = IndexDefinitionStorage.GetIndexDefinition(index.Id);
+		                    index.LastQueryTimestamp = IndexStorage.GetLastQueryTime(index.Id);
+		                    index.Performance = IndexStorage.GetIndexingPerformance(index.Id);
+		                    index.IsOnRam = IndexStorage.IndexOnRam(index.Id);
 		                    if (indexDefinition != null)
 			                    index.LockMode = indexDefinition.LockMode;
 		                    index.ForEntityName = IndexDefinitionStorage.GetViewGenerator(index.Name).ForEntityNames.ToList();
@@ -1037,10 +1037,10 @@ namespace Raven.Database
                                         continue;
                                 }
 
-                                string indexNameCopy = indexName;
-                                var task = actions.GetTask(x => x.Index == indexNameCopy, new RemoveFromIndexTask
+                                var instance = IndexDefinitionStorage.GetIndexDefinition(indexName);
+                                var task = actions.GetTask(x => x.Index == instance.IndexId, new RemoveFromIndexTask
                                 {
-                                    Index = indexNameCopy
+                                    Index = instance.IndexId
                                 });
                                 task.Keys.Add(key);
                             }
@@ -1158,7 +1158,7 @@ namespace Raven.Database
                 return name; // no op for the same transformer
 
             IndexDefinitionStorage.CreateAndPersistTransform(definition);
-            IndexDefinitionStorage.AddTransform(name, definition);
+            IndexDefinitionStorage.AddTransform(definition.IndexId, definition);
 
             return name;
         }
@@ -1195,7 +1195,7 @@ namespace Raven.Database
                     return name;
                 case IndexCreationOptions.Update:
                     // ensure that the code can compile
-                    new DynamicViewCompiler(name, definition, Extensions, IndexDefinitionStorage.IndexDefinitionsPath, Configuration).GenerateInstance();
+                    new DynamicViewCompiler(definition.Name, definition, Extensions, IndexDefinitionStorage.IndexDefinitionsPath, Configuration).GenerateInstance();
                     DeleteIndex(name);
                     break;
             }
@@ -1209,14 +1209,14 @@ namespace Raven.Database
 
             TransactionalStorage.Batch(actions =>
             {
-                actions.Indexing.AddIndex(name, definition.IsMapReduce);
+                actions.Indexing.AddIndex(definition.IndexId, definition.IsMapReduce);
                 workContext.ShouldNotifyAboutWork(() => "PUT INDEX " + name);
             });
 
             // The act of adding it here make it visible to other threads
             // we have to do it in this way so first we prepare all the elements of the 
             // index, then we add it to the storage in a way that make it public
-            IndexDefinitionStorage.AddIndex(name, definition);
+            IndexDefinitionStorage.AddIndex(definition.IndexId, definition);
 
             InvokeSuggestionIndexing(name, definition);
 
@@ -1258,6 +1258,7 @@ namespace Raven.Database
         private IndexCreationOptions FindIndexCreationOptions(IndexDefinition definition, ref string name)
         {
 	        definition.Name = name;
+            definition.IndexId = IndexDefinitionStorage.NextIndexId();
             definition.RemoveDefaultValues();
             IndexDefinitionStorage.ResolveAnalyzers(definition);
             var findIndexCreationOptions = IndexDefinitionStorage.FindIndexCreationOptions(definition);
@@ -1274,8 +1275,14 @@ namespace Raven.Database
 
         public QueryResultWithIncludes Query(string index, IndexQuery query, Action<QueryHeaderInformation> headerInfo, Action<RavenJObject> onResult)
         {
+            index = IndexDefinitionStorage.FixupIndexName(index);
             var queryStat = AddToCurrentlyRunningQueryList(index, query);
+            var highlightings = new Dictionary<string, Dictionary<string, string[]>>();
             try
+	        var scoreExplanations = new Dictionary<string, string>();
+            var definition = IndexDefinitionStorage.GetIndexDefinition(index);
+
+            Func<IndexQueryResult, object> tryRecordHighlightingAndScoreExplanation = queryResult =>
             {
 
 	            index = index != null ? index.Trim() : null;
@@ -1499,6 +1506,9 @@ namespace Raven.Database
         {
             using (IndexDefinitionStorage.TryRemoveIndexContext())
             {
+                var instance = IndexDefinitionStorage.GetIndexDefinition(name);
+                if (instance == null) return;
+                name = IndexDefinitionStorage.FixupIndexName(name);
                 IndexDefinitionStorage.RemoveIndex(name);
                 IndexStorage.DeleteIndex(name);
                 //we may run into a conflict when trying to delete if the index is currently
@@ -1510,8 +1520,7 @@ namespace Raven.Database
                     {
                         TransactionalStorage.Batch(action =>
                         {
-                            action.Indexing.DeleteIndex(name);
-
+                            action.Indexing.DeleteIndex(instance.IndexId);
                             workContext.ShouldNotifyAboutWork(() => "DELETE INDEX " + name);
                         });
 
@@ -2334,14 +2343,14 @@ namespace Raven.Database
             {
 				var indexInstance = IndexStorage.GetIndexInstance(indexName);
 	            isStale = (indexInstance != null && indexInstance.IsMapIndexingInProgress) ||
-	                      accessor.Staleness.IsIndexStale(indexName, null, null);
+	                      accessor.Staleness.IsIndexStale(indexInstance.indexId, null, null);
                 lastDocEtag = accessor.Staleness.GetMostRecentDocumentEtag();
-                var indexStats = accessor.Indexing.GetIndexStats(indexName);
+                var indexStats = accessor.Indexing.GetIndexStats(indexInstance.indexId);
                 if (indexStats != null)
                 {
                     lastReducedEtag = indexStats.LastReducedEtag;
                 }
-                touchCount = accessor.Staleness.GetIndexTouchCount(indexName);
+                touchCount = accessor.Staleness.GetIndexTouchCount(indexInstance.indexId);
             });
